Abstraction and 2d text-based example

+++ src/main/java/Engine/Node.java
@@ -11,19 +11,19 @@ package Engine;
 import java.util.Objects;
 
 public class Node {
-    private final int x;
-    private final int y;
+    private final double x;
+    private final double y;
     
-    public Node(int x, int y) {
+    public Node(double x, double y) {
         this.x = x;
         this.y = y;
     }
     
-    public int getX() {
+    public double getX() {
         return x;
     }
     
-    public int getY() {
+    public double getY() {
         return y;
     }
